Add files and folders.
