+++ backend/src/config.py
@@ -1,0 +1,27 @@
+import sentry_sdk
+from apscheduler.schedulers.background import BackgroundScheduler
+from fastapi.middleware.cors import CORSMiddleware
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+# Sentry config
+sentry_sdk.init(
+    dsn="https://4001ffe917ccb261aa0e0c34026dc343@o4505702629834752.ingest.us.sentry.io/4507694792704000",
+    traces_sample_rate=1.0,
+    profiles_sample_rate=1.0,
+)
+
+# Database engine
+engine = create_engine("sqlite:///news_database.db", echo=True)
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+# Scheduler
+bgs = BackgroundScheduler()
+
+# CORS middleware settings
+CORS_OPTIONS = {
+    "allow_origins": ["http://localhost:8080"],
+    "allow_credentials": True,
+    "allow_methods": ["*"],
+    "allow_headers": ["*"],
+}
