+++ backend/src/model.py
@@ -1,0 +1,33 @@
+from sqlalchemy import Column, Integer, String, Text, ForeignKey, Table
+from sqlalchemy.orm import relationship
+from .database import Base
+
+# Association table for user and news articles
+user_news_association_table = Table(
+    "user_news_upvotes",
+    Base.metadata,
+    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
+    Column("news_articles_id", Integer, ForeignKey("news_articles.id"), primary_key=True),
+)
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    username = Column(String(50), unique=True, nullable=False)
+    hashed_password = Column(String(200), nullable=False)
+    upvoted_news = relationship(
+        "NewsArticle", secondary=user_news_association_table, back_populates="upvoted_by_users"
+    )
+
+class NewsArticle(Base):
+    __tablename__ = "news_articles"
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    url = Column(String, unique=True, nullable=False)
+    title = Column(String, nullable=False)
+    time = Column(String, nullable=False)
+    content = Column(Text, nullable=False)
+    summary = Column(Text, nullable=False)
+    reason = Column(Text, nullable=False)
+    upvoted_by_users = relationship(
+        "User", secondary=user_news_association_table, back_populates="upvoted_news"
+    )
